% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_htmls.R
\name{build_temp_rmd}
\alias{build_temp_rmd}
\title{Generate a temporary .Rmd file}
\usage{
build_temp_rmd(orig_rmd_path, temp_rmd_path)
}
\arguments{
\item{orig_rmd_path}{character of length = 1
A path to an original .Rmd file.}

\item{temp_rmd_path}{character of length = 1
A name ("--" is a part of those names) of a temporary .Rmd file.}
}
\value{
Temporarily saved .Rmd files in directory "analysis".
}
\description{
Generate a temporary (helping) .Rmd file from its original .Rmd file
and temporarily save it into directory "analysis".
This temporarily saved .Rmd file will be used to generate final .html file and
will be deleted at the end of function \code{render_to_htmls} after final .html file is prepared.
This function is called only from \code{render_to_htmls} so its input variables have no default values.
}
