% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.rmd.R
\name{generate.rmd}
\alias{generate.rmd}
\title{Generate temporary .Rmd files}
\usage{
generate.rmd(dir = NULL, path = NULL, temp_file = NULL)
}
\arguments{
\item{dir}{path to a directory, under a main workflowr subdirectory, where an original .Rmd file is saved.}

\item{path}{character (default: NULL). Path to an original .Rmd file saved in a dir's subdirectory.}

\item{temp_file}{character (default: NULL).
Name ("--" is usually part of it's name) of a temporary file that will be saved into directory "analysis".
This temporary file is generated from its original .Rmd file specified in path, then it will be deleted within generate.subdir.}
}
\description{
This is one of functions serving as optional extension for workflowr package in order to be able to generate .html files from .Rmd files saved in subdirectories and not only in directory "analysis".
generate.rmd generates .Rmd files from original .Rmd files that are saved in subdirectories and save them into directory "analysis".
These generated .Rmd files are meant to be temporary and they will be deleted after generate.subdir generates final .html files.
}
\examples{
cat_function()
used literature: _preparationSubfolders.R
}
\keyword{workflowr}
