% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-html.R
\name{generate_html}
\alias{generate_html}
\title{Render .html files from their original .Rmd files stored in subdirectories}
\usage{
generate_html(
  dir = "code-Rmd",
  only_subdirs = NULL,
  orig_rmd_pattern = NULL,
  commit = F
)
}
\arguments{
\item{dir}{character (default: "code-Rmd").
Path to a directory, under a main workflowr subdirectory, where original Rmd files are saved.}

\item{only_subdirs}{character (default: NULL). It's case insensitive.
If only_subdirs == NULL then all subdirectories and files within directory in input parameter dir are processed, otherwise only files in subdirectories in this input parameter only_subdirs are processed.
If only_subdirs != NULL then it's a vector of subdirectories in directory specified in input parameter dir.
Examples: only_subdirs = NULL; only_subdirs == c("subdir1", "subdir.Rmd")}

\item{orig_rmd_pattern}{character (default: NULL).
If orig_rmd_pattern == NULL then search for all files in directories set by dir or only_subdirs.
Vector of paths to original .Rmd files. These file paths start with a name of the 1st subdirectory of a directory specified in variable "dir".
Example when directories subPagesX are saved in directory dir = "code-Rmd":
file_path = c("subPages2/testPrint1.Rmd", "subPages3/testPrint2.Rmd")
file_path = c("subPages2\\testPrint1.Rmd", "subPages3\\testPrint2.Rmd")}

\item{commit}{character (default: FALSE).
commit = TRUE creates a separate commit of temporary .Rmd files (temporary saved in directory "analysis").
Suggestion: Use commit = TRUE only after your original .Rmd files saved in subdirectories are tested properly and so are completely ready, otherwise you could have pointlessly many commits.}
}
\value{
Final .html file from its original .Rmd file saved in a subdirectory.
}
\description{
Similarly as for workflowr package, when specifying path to directories or files in function generate_html(), you can start by typing directory name, so for example, you can write "analysis" instead of "./analysis".
Processed steps:
\enumerate{
\item Check existence of .Rmd file in a chosen subdirectory.
}
\itemize{
\item If it doesn't exist, inform about it and stop processing.
}
\enumerate{
\item Prepare temporary file names by substitution  of "/" in file path by "‐‐".
\item Function generate_rmd() is used to generate temporary .Rmd files from their original .Rmd files that are saved in subdirectories and save them into directory "analysis".
\item If commit = TRUE, create a commit of these temporary .Rmd files with text "separate commit of temporary .Rmd files".
\item Generate .html files from temporary .Rmd files.
\itemize{
\item So the resulting .html files are associated with their original .Rmd files (temporary .Rmd files are simply product of a helping step).
}
\item Delete temporary .Rmd files from directory "analysis".
}
}
\examples{
\dontrun{
  generate_html()
  generate_html("code-Rmd", c("subPages1/testPrint1.Rmd", "subPages2/testPrint2.Rmd"), T)
  generate_html("code-Rmd", c("subPages1\\\\testPrint1.Rmd", "subPages2\\\\testPrint2.Rmd"), F)
}
}
\keyword{subdirectory}
\keyword{workflowr,}
