% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-htmls.R
\name{generate_htmls}
\alias{generate_htmls}
\title{Generate .html files from their associate .Rmd files}
\usage{
generate_htmls(dir_path = "code-rmd", subdirs = T, patterns = NULL, commit = F)
}
\arguments{
\item{dir_path}{character of length = 1 (default: "code-rmd").
Path to a subdirectory, under a main workflowr directory, where .Rmd files for rendering are saved.
A directory name can be first instead of using "./" etc. (e.g. "code-rmd" instead of "./code-rmd").
Examples:
dir_path = "code-rmd"
dir_path = c("code-rmd/subdir1\\subdir2")}

\item{subdirs}{logical of length = 1 (default: TRUE)
If TRUE, file listing will recurse into directories in parameter dir.
If FALSE, file listing will be only directly from directories in parameter dir.}

\item{patterns}{character of length > 0 or NULL (default: NULL)
A character vector of paths to .Rmd files for rendering.
If NULL, process all .Rmd files based values in parameters dir and subdirs.
If not NULL, process files matching a regular expression.
Examples:
patterns = "^.\emph{page.}.[  R , r ]md$")
patterns = c("page1.Rmd", ".\emph{page2.}.Rmd")}

\item{commit}{character (default: FALSE)
If TRUE, a separate commit of temporary .Rmd files (temporary saved in "analysis") is created.
It's suggested to use "commit = TRUE" only after original .Rmd files saved
in subdirectories are tested properly and so are completely ready,
otherwise there could be uselessly many commits.}
}
\value{
Final .html files from their original .Rmd files saved in subdirectories.
}
\description{
Process only those .Rmd files (based on input parameters) that meet required criteria.
}
\examples{
\dontrun{
  generate_htmls()
  generate_htmls(dir_path = c("code-rmd\\\\subdir"), subdirs = F)
  generate_htmls("code-rmd/subdir", T, c("file1.Rmd", "-.*.[ R , r ]md"))
}
}
\keyword{subdirectory}
\keyword{workflowr,}
