% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rmd.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Generate temporary .Rmd files}
\usage{
generate_rmd(dirs = "code-rmd", orig_rmd_path = NULL, temp_rmd_path = NULL)
}
\arguments{
\item{dirs}{character (default: "code-rmd").
Paths to directories, under a main workflowr directory, where original .Rmd files are saved.
Examples:
dirs = "code-rmd"
dirs = c("code-rmd/subpage", "code-rmd/subpage1\\subpage2")}

\item{orig_rmd_path}{character (default: NULL).
Path to an original .Rmd file. Path ignores dirs (see the 1st parameter).}

\item{temp_rmd_path}{character (default: NULL).
Name ("--" is usually part of it's name) of a temporary .Rmd file that will be temporarily saved into directory "analysis".
This temporary file is generated from its original .Rmd file specified in path_Rmd, then it will be deleted within \code{\link{generate_html}}.}
}
\value{
Temporary .Rmd file, from its original .Rmd file saved in a subdirectory, used to generate final .html file.
}
\description{
Generate temporary .Rmd files from their original .Rmd files,
usually saved in directory "code" and its subdirectories,
and save them into directory "analysis".
These generated .Rmd files will be deleted at the end of function \code{\link{generate_html}}
after final .html files are rendered.
}
\keyword{subdirectory}
\keyword{workflowr,}
