% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-to-htmls.R
\name{render_to_htmls}
\alias{render_to_htmls}
\title{Render .Rmd files to .html files}
\usage{
render_to_htmls(orig_rmd_paths, commit)
}
\arguments{
\item{orig_rmd_paths}{character
Paths to original (associated) .Rmd files.}

\item{commit}{character
If TRUE, a separate commit of temporary .Rmd files (temporary saved in "analysis") is created.
It's suggested to use "commit = TRUE" only after original .Rmd files saved
in subdirectories are tested properly and so are completely ready,
otherwise there could be uselessly many commits.}
}
\value{
Final .html files from their original .Rmd files saved in subdirectories.
}
\description{
Render only those .Rmd files (based on input parameters) that meet required criteria.
}
\examples{
\dontrun{
  render_to_htmls(orig_rmd_paths, T)
}
}
\keyword{subdirectory}
\keyword{workflowr,}
