% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-htmls.R
\name{build_htmls}
\alias{build_htmls}
\title{Generate .html files from their original .Rmd files}
\usage{
build_htmls(dir_path = "code-rmd", subdirs = T, patterns = NULL, commit = F)
}
\arguments{
\item{dir_path}{character of length = 1 (default: "code-rmd").
Path to directory "code-rmd" or its subdirectories where .Rmd files
for rendering to .html files are saved. This path is within
a current working workflowr directory. Following examples also show that
slashes can be used but they are not necessary at the beginning and end of path.
Examples: "code-rmd", c("code-rmd/subdir1\\subdir2"),
"./code-rmd/subdir/", "code-rmd/eToro1\\", "D:/project/code-rmd/subdir"}

\item{subdirs}{logical of length = 1 (default: TRUE)
If TRUE, file listing will also recurse into directories in parameter dir_path.
If FALSE, file listing will be only directly from a directory in parameter dir_path.}

\item{patterns}{character of length > 0 or NULL (default: NULL)
A character vector of paths to .Rmd files for rendering.
If NULL, process all .Rmd files based values in parameters dir_path and subdirs.
If not NULL, process files matching used regular expressions which
always have to end with ".rmd", ".Rmd", ".rmd$", ".Rmd$" or
a relevant regular expression that after evaluation point to
extension ".rmd" or ".Rmd" ("." is also required). This is made
in accordance to behavior of package "workflowr"
which allows only ".rmd" or ".Rmd" extensions.
Examples: "^.\emph{page.}.[  R , r ]md$"), c("page1.Rmd", ".\emph{page2.}.Rmd")}

\item{commit}{character (default: FALSE)
If TRUE, a separate commit of temporary .Rmd files (temporary saved in "analysis") is created.
Consider to use "commit = TRUE" only after original .Rmd files
are properly completed otherwise there could be uselessly many commits.}
}
\value{
Final .html files from their original .Rmd files.
}
\description{
Process only those .Rmd files (based on input parameters) that meet required criteria.
Use parameters "dir_path", "subdirs" and "patterns" to determine which .rmd or .Rmd files
will be rendered to .html files.
}
\examples{
\dontrun{
  build_htmls()
  build_htmls(dir_path = c("code-rmd\\\\subdir"), subdirs = F)
  build_htmls("code-rmd/subdir", T, c("file1.Rmd", "-.*.[ R , r ]md"), T)
}
}
