% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-htmls.R
\name{build_htmls}
\alias{build_htmls}
\title{Build HTML files}
\usage{
build_htmls(dir_path = "code-rmd", subdirs = T, patterns = NULL, commit = F)
}
\arguments{
\item{dir_path}{character (length = 1; default: "code-rmd").
A path directly to directory "code-rmd" or one of its subdirectories with
R Markdown files specified in input parameter \code{patterns}.
Only a path within a current working workflowr directory is allowed.}

\item{subdirs}{logical (length = 1; default: TRUE).
If \code{FALSE}, file listing will only be directly from a directory in parameter \code{dir_path}.
If \code{TRUE}, file listing will also recurse into subdirectories in parameter \code{dir_path}.}

\item{patterns}{character (length > 0; default: NULL).
If \code{not NULL}, one or more regular expressions is expected.
If \code{NULL}, regular expression "\\.(r|R)md$" is automatically used.}

\item{commit}{character (default: FALSE).
If \code{FALSE}, nothing from actions for \code{TRUE} happens.
If \code{TRUE}, a separate commit of temporary (automatically saved and later deleted)
R Markdown files saved in directory "analysis" is made and built HTML files contain
checked line "R Markdown file" in WORKFLOWR button in tab Checks.
Consider to use \code{commit = TRUE} only after original R Markdown files
are completely finished, otherwise uselessly many commits are made.}
}
\value{
HTML files from their associated original R Markdown files.
}
\description{
Multiple R Markdown files saved in directory "code-rmd" and its subdirectories
can be specified for rendering to HTML files.
The HTML files are always built in the order their associated R Markdown files are listed.
}
\examples{
\dontrun{

# Build HTML files from all .(r|R)md files
  in "code-rmd" and its subdirectories
build_htmls()
# Build a single file
build_htmls("code-rmd/subdir/sub1", F, "file.Rmd")
# Build multiple files (not equivalent examples)
build_htmls(patterns = "(?i)RMD")
build_htmls("code-rmd/subdir", F)
build_htmls("code-rmd/subdir", patterns = "\\\\.(r|R)md$")
build_htmls("code-rmd\\\\subdir", T, c("-.*.[rR]md", "file"), T)
}
}
