% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-htmls.R
\name{build_temp_rmd}
\alias{build_temp_rmd}
\title{Generate a temporary R Markdown file}
\usage{
build_temp_rmd(temp_c_rmd_path, temp_a_rmd_path)
}
\arguments{
\item{temp_c_rmd_path}{character (length = 1).
A path to an R Markdown file temporarily copied from its original R Markdown file
directly to directory "code-rmd".
This file will be deleted at the end of function \code{render_to_htmls}.}

\item{temp_a_rmd_path}{character (length = 1).
A path to an R Markdown file that will be temporarily saved in directory "analysis".
This file will be deleted at the end of function \code{render_to_htmls}.}
}
\value{
An R Markdown file temporarily saved in directory "analysis".
}
\description{
Generate a temporary (helping) R Markdown file, that will be used to generate final HTML file,
from its original R Markdown file and temporarily save it into directory "analysis".
}
