% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-orig-Rmd-path.R
\name{create_orig_Rmd_path}
\alias{create_orig_Rmd_path}
\title{Create paths to original .Rmd files for future rendering}
\usage{
create_orig_Rmd_path(
  dir = "code-Rmd",
  only_subdirs = NULL,
  orig_Rmd_pattern = NULL
)
}
\arguments{
\item{dir}{character (default: "code-Rmd").
Path to a directory, under a main workflowr subdirectory, where original Rmd files are saved.}

\item{only_subdirs}{character (default: NULL). It's case insensitive.
If only_subdirs == NULL then all subdirectories and files within directory in input parameter dir are processed, otherwise only files in subdirectories in this input parameter only_subdirs are processed.
If only_subdirs != NULL then it's a vector of subdirectories in directory specified in input parameter dir.
Examples: only_subdirs = NULL; only_subdirs == c("subdir1", "subdir.Rmd")}

\item{orig_Rmd_pattern}{character (default: NULL).
Vector of paths to original .Rmd files. These file paths start with a name of the 1st subdirectory of a directory specified in variable "dir".
Example when directories subPagesX are saved in directory dir = "code-Rmd":
file_path = c("subPages2/testPrint1.Rmd", "subPages3/testPrint2.Rmd")
file_path = c("subPages2\\testPrint1.Rmd", "subPages3\\testPrint2.Rmd")}
}
\value{
None, but stop processing if any of checks fails.
}
\description{
Check if rendering .html files from .Rmd files from subdirectories is possible and
in such case, create paths to original .Rmd files for future rendering otherwise stop processing.
}
\examples{
\dontrun{
  initial_checks(dir, path_orig_Rmd)
}
}
\keyword{subdirectory}
\keyword{workflowr,}
