% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-orig-rmd-path.R
\name{create_orig_rmd_path}
\alias{create_orig_rmd_path}
\title{Create paths to original .Rmd files for future rendering}
\usage{
create_orig_rmd_path(dirs, subdirs, orig_rmd_pattern)
}
\arguments{
\item{dirs}{character
Paths to directories, under a main workflowr directory, where original .Rmd files are saved.
Examples:
dirs = "code-rmd"
dirs = c("code-rmd/subpage", "code-rmd/subpage1\\subpage2")}

\item{subdirs}{logical
If TRUE, file listing will also recurse into directories in parameter dir.
If FALSE, file listing will be only directly from directories in parameter dir.}

\item{orig_rmd_pattern}{character
Vector of paths to original .Rmd files.
If NULL, process all .Rmd files based values in parameters dir and subdirs.
If not NULL, process files matching written regular expression.
Examples:
orig_rmd_pattern = "^.\emph{page.}.[  R , r ]md$")
orig_rmd_pattern = c("page1.Rmd", ".\emph{page2.}.Rmd")}
}
\value{
Character vector "orig_rmd_path" but stop processing if no file meets criteria.
}
\description{
Create paths to original .Rmd files for future rendering into .html.
If no file for rendering is found, processing ends.
This function is meant to be called only from function \code{\link{generate_html}} so
it will be not exported therefore input variables have no default values.
}
\examples{
\dontrun{
  create_orig_rmd_path(dirs = "code-rmd", subdirs = T, orig_rmd_pattern = ".*page.*.(R|r)md$")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
