% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmd_paths.R
\name{create_rmd_paths}
\alias{create_rmd_paths}
\title{Create paths to .Rmd files for future rendering}
\usage{
create_rmd_paths(dir_path, subdirs, patterns)
}
\arguments{
\item{dir_path}{character of length = 1
Path to a subdirectory, under a main workflowr directory, where .Rmd files for rendering are saved.
A directory name can be first instead of using "./" etc. (e.g. "code-rmd" instead of "./code-rmd").
Examples:
dir_path = "code-rmd"
dir_path = c("code-rmd/subdir1\\subdir2")}

\item{subdirs}{logical of length = 1
If TRUE, file listing will recurse into directories in parameter dir.
If FALSE, file listing will be only directly from directories in parameter dir.}

\item{patterns}{character of length > 0 or NULL
A character vector of paths to .Rmd files for rendering.
If NULL, process all .Rmd files based values in parameters dir and subdirs.
If not NULL, process files matching a regular expression.
Examples:
patterns = "^.\emph{page.}.[  R , r ]md$")
patterns = c("page1.Rmd", ".\emph{page2.}.Rmd")}
}
\value{
FALSE if no file meets criteria, a character vector if it's created.
}
\description{
Create paths to original .Rmd files intended for future rendering into .html files.
If input parameters point to
\itemize{
\item a .Rmd file path that doesn't exist then such file is ignored (not rendered).
\item a .Rmd file path that exist then such file is rendered in \code{\link{render_html}} using wflow_build().
If no file for rendering is found, processing ends.
This function is called only from function \code{\link{render_html}} therefore
it's not exported and its input variables have no default values.
}
}
\examples{
\dontrun{
  create_rmd_paths(dir_path = "code-rmd", subdirs = T, patterns = ".*page.*.(R|r)md$")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
