% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_htmls.R
\name{create_rmd_paths}
\alias{create_rmd_paths}
\title{Create paths to original .Rmd files}
\usage{
create_rmd_paths(dir_path, subdirs, patterns)
}
\arguments{
\item{dir_path}{see \code{build_htmls}}

\item{subdirs}{see \code{build_htmls}}

\item{patterns}{see \code{build_htmls}}
}
\value{
A character vector with original .Rmd file paths if at least one .Rmd file meets criteria.
Nothing and stop processing if no file meets criteria.
}
\description{
Create paths to original .Rmd files intended for future rendering into .html files.
If input parameters point to a .Rmd file path that
\itemize{
\item doesn't exist then such file will be not rendered.
\item exist then such file will be rendered in \code{render-to-htmls} using wflow_build().
If no .Rmd file for rendering is found, processing ends.
This function is called only from \code{build_htmls} so its input variables have no default values.
}
}
