% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rmd_paths.R
\name{create_rmd_paths}
\alias{create_rmd_paths}
\title{Create paths to .Rmd files for future rendering}
\usage{
create_rmd_paths(dir_path, subdirs, patterns)
}
\arguments{
\item{dir_path}{character
Path to subdirectory, under a main workflowr directory, where original .Rmd files are saved.
It can be only of length = 1.
Examples:
dir_path = "code-rmd"
dir_path = c("code-rmd/subdir1\\subdir2")}

\item{subdirs}{logical
It can be only of length = 1.
If TRUE, file listing will also recurse into directories in parameter dir.
If FALSE, file listing will be only directly from directories in parameter dir.}

\item{patterns}{character
Vector of paths to original .Rmd files.
If NULL, process all .Rmd files based values in parameters dir and subdirs.
If not NULL, process files matching written regular expression.
Examples:
patterns = "^.\emph{page.}.[  R , r ]md$")
patterns = c("page1.Rmd", ".\emph{page2.}.Rmd")}
}
\value{
Character vector "orig_rmd_path" but stop processing if no file meets criteria.
}
\description{
Create paths to (usually) original .Rmd files, based on input parameters,
for future rendering into .html.
If no file for rendering is found, processing ends.
This function is meant to be called only from function \code{\link{render_html}} so
it will be not exported therefore input variables have no default values.
}
\examples{
\dontrun{
  create_rmd_paths(dir_path = "code-rmd", subdirs = T, patterns = ".*page.*.(R|r)md$")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
