% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-htmls.R
\name{create_rmd_paths}
\alias{create_rmd_paths}
\title{Create paths}
\usage{
create_rmd_paths(dir_path, subdirs, patterns)
}
\arguments{
\item{dir_path}{character (length = 1; default: "code-rmd").
A path directly to directory "code-rmd" or one of its subdirectories with
R Markdown files specified in input parameter \code{patterns}.
Only a path within a current working workflowr directory is allowed.}

\item{subdirs}{logical (length = 1; default: TRUE).
If \code{FALSE}, file listing will only be directly from a directory in parameter \code{dir_path}.
If \code{TRUE}, file listing will also recurse into subdirectories in parameter \code{dir_path}.}

\item{patterns}{character (length > 0; default: NULL).
If \code{not NULL}, one or more regular expressions that only allow files
with the extension .rmd or .Rmd ("." is also required) is expected.
If \code{NULL}, regular expression "\\.(r|R)md$" is used.}
}
\value{
Nothing if no file meets criteria.
A character vector with original R Markdown file paths,
if at least one R Markdown file meets criteria.
}
\description{
Create paths to original R Markdown files intended for future rendering to HTML files.
If no .Rmd file for rendering is found, processing ends.
If input parameters refer to an R Markdown file path that
\itemize{
\item doesn't exist then such file will be not rendered.
\item exist then such file will be rendered in \code{render-to-htmls} using wflow_build().
}
}
