% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-html.R
\name{generate_html}
\alias{generate_html}
\title{Render .html files from their original .Rmd files stored in subdirectories}
\usage{
generate_html(dir = "codeRmd", file_path = NULL, commit = F)
}
\arguments{
\item{dir}{character (default: "codeRmd").
Path to a directory, under a main workflowr subdirectory, where original Rmd files are saved.}

\item{file_path}{character (default: NULL).
Vector of paths to original .Rmd files. These file paths start with a name of the 1st subdirectory of a directory specified in variable "dir".
Example when directories subPagesX are saved in folder dir = "codeRmd": file_path = c("subPages2/testPrint2.Rmd", "subPages3/testPrint3.Rmd")}

\item{commit}{character (default: FALSE).
commit = TRUE creates a separate commit of temporary .Rmd files (temporary saved in directory "analysis").
Suggestion: Use commit = TRUE only after your original .Rmd files saved in subdirectories are tested properly and so are completely ready, otherwise you could have pointlessly many commits.}
}
\value{
<return>
}
\description{
Similarly as for workflowr package, when specifying path to folders or files in function generate_html(), you can start by typing folder name, so for example, you can write "analysis" instead of "./analysis".
Processed steps:
\enumerate{
\item Check existence of .Rmd file in a chosen subdirectory.
}
\itemize{
\item If it doesn't exist, inform about it and stop processing.
}
\enumerate{
\item Prepare temporary file names by substitution  of "/" in file path by "‐‐".
\item Function generate_rmd() is used to generate temporary .Rmd files from their original .Rmd files that are saved in subdirectories and save them into directory "analysis".
\item If commit = TRUE, create a commit of these temporary .Rmd files with text "separate commit of temporary .Rmd files".
\item Generate .html files from temporary .Rmd files.
\itemize{
\item So the resulting .html files are associated with their original .Rmd files (temporary .Rmd files are simply product of a helping step).
}
\item Delete temporary .Rmd files from directory "analysis".
}
}
\examples{
\dontrun{
  generate_html()
  generate_html("codeRmd", c("subPages1/testPrint1.Rmd", "subPages2/testPrint2.Rmd"), T)
}
}
\keyword{subdirectory}
\keyword{workflowr,}
