% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rmd.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Generate a temporary .Rmd file}
\usage{
generate_rmd(orig_rmd_path, temp_rmd_path)
}
\arguments{
\item{orig_rmd_path}{character of length = 1
A path to original .Rmd files.}

\item{temp_rmd_path}{character of length = 1
Names ("--" is a part of those names) of temporary .Rmd files.}
}
\value{
Temporarily saved .Rmd files.
}
\description{
Generate a temporary (helping) .Rmd file, from its original .Rmd file
specified using input parameters "dir_path", "subdirs" and "patterns" of \code{\link{render_html}},
and temporarily save it into directory "analysis".
This temporarily saved .Rmd file will be used to generate final .html file and
will be deleted at the end of function \code{\link{render_html}} after final .html file is prepared.
This function requires that both input parameters are of length = 1 but no checks for that
is made directly inside this function because this is ensured using mapply() in \code{\link{render_html}}.
This function is called only from function \code{\link{render_html}} therefore
it's not exported and its input variables have no default values.
}
\examples{
\dontrun{
  generate_rmd("code-rmd/subdir/testfile.Rmd", "analysis/subdir--testfile.Rmd")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
