% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rmd.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Generate temporary .Rmd files}
\usage{
generate_rmd(orig_rmd_path, temp_rmd_path)
}
\arguments{
\item{orig_rmd_path}{character
Path to an original .Rmd file saved in workflowr subdirectory.}

\item{temp_rmd_path}{character
Name ("--" is usually part of it's name) of a temporary .Rmd file that will be
temporarily saved into directory "analysis".
This temporary file is generated from its original .Rmd file specified in path_Rmd,
then it will be deleted within \code{\link{render_html}}.}
}
\value{
Temporary .Rmd file, from its original .Rmd file saved in a subdirectory,
used to generate final .html file.
}
\description{
Generate temporary .Rmd files from their original .Rmd files,
usually saved in directory "code" and its subdirectories,
and save them into directory "analysis".
These generated .Rmd files will be deleted at the end of function \code{\link{render_html}}
after final .html files are rendered.
This function is meant to be called only from function \code{\link{render_html}} so
it will be not exported therefore input variables have no default values.
}
\examples{
\dontrun{
  generate_rmd("code-rmd/subdir/testfile.Rmd", "analysis/subdir--testfile.Rmd")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
