% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rmd.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Generate temporary .Rmd files}
\usage{
generate_rmd(dir = "code-Rmd", orig_rmd_path = NULL, temp_rmd_path = NULL)
}
\arguments{
\item{dir}{character (default: "code-Rmd").
Path to a directory, under a main workflowr subdirectory, where original Rmd files are saved.}

\item{orig_rmd_path}{character (default: NULL).
Path to an original .Rmd file. Path ignores dir (see the 1st parameter).}

\item{temp_rmd_path}{character (default: NULL).
Name ("--" is usually part of it's name) of a temporary .Rmd file that will be temporarily saved into directory "analysis".
This temporary file is generated from its original .Rmd file specified in path_Rmd, then it will be deleted within \code{\link{generate_html}}.}
}
\value{
Temporary .Rmd file, from its original .Rmd file saved in a subdirectory, used to generate final .html file.
}
\description{
It generates .Rmd files from their original .Rmd files that are saved in subdirectories and saves them into directory "analysis".
These generated .Rmd files are meant to be temporary and they will be deleted at the end of function \code{\link{generate_html}} after final .html files are rendered.
}
\examples{
\dontrun{
  generate_rmd("code-Rmd", "subPages1/testPrint1.Rmd", "subPages1--testPrint1.Rmd")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
