% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rmd.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Generate temporary .Rmd files}
\usage{
generate_rmd(dir = "codeRmd", path = NULL, temp_file = NULL)
}
\arguments{
\item{dir}{character (default: "codeRmd") path to a directory, under a main workflowr subdirectory, where an original Rmd file is saved.}

\item{path}{character (default: NULL). Path to an original .Rmd file saved in a dir or in a dir's subdirectory.}

\item{temp_file}{character (default: NULL).
Name ("--" is usually part of it's name) of a temporary .Rmd file that will be saved into directory "analysis".
This temporary file is generated from its original Rmd file specified in path, then it will be deleted within generate.subdir.}
}
\value{
<return>
}
\description{
This is one of functions serving as optional extension for workflowr package in order to be able to generate .html files from .Rmd files saved in subdirectories and not only in directory "analysis".
\code{generate_rmd} generates .Rmd files from original .Rmd files that are saved in subdirectories and save them into directory "analysis".
These generated .Rmd files are meant to be temporary and they will be deleted after \code{\link{generate_subdir}} generates final .html files.
If you want to use this function in other way than call it from \code{\link{generate_subdir}} then you have to set a correct workflowr project to be your working directory,
e.g. by running: setwd("D:/Cloud/Sync/Study/Programming/R/Work/Projects/GitHub/workflowrSubfolders")
example: #' generate_rmd("codeRmd", "subPages1/testPrint1.Rmd", "subPages1--testPrint1.Rmd")
}
\examples{
print("a")
}
\keyword{subdirectory}
\keyword{workflowr,}
