% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rmd.R
\name{generate_rmd}
\alias{generate_rmd}
\title{Generate temporary .Rmd files}
\usage{
generate_rmd(dir = "codeRmd", file_path = NULL, temp_file = NULL)
}
\arguments{
\item{dir}{character (default: "codeRmd").
Path to a directory, under a main workflowr subdirectory, where original Rmd files are saved.}

\item{file_path}{character (default: NULL).
Path to an original .Rmd file saved in a dir or in a dir's subdirectory.}

\item{temp_file}{character (default: NULL).
Name ("--" is usually part of it's name) of a temporary .Rmd file that will be saved into directory "analysis".
This temporary file is generated from its original Rmd file specified in path, then it will be deleted within generate.subdir.}
}
\value{
<return>
}
\description{
It generates .Rmd files from their original .Rmd files that are saved in subdirectories and saves them into directory "analysis".
These generated .Rmd files are meant to be temporary and they will be deleted at the end of function \code{\link{generate_subdir}} after final .html files are rendered.
}
\examples{
\dontrun{
  generate_rmd("codeRmd", "subPages1/testPrint1.Rmd", "subPages1--testPrint1.Rmd")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
