% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-htmls.R
\name{initial_checks}
\alias{initial_checks}
\title{Make initial checks and preparations}
\usage{
initial_checks(dir_path, subdirs, patterns)
}
\arguments{
\item{dir_path}{character (length = 1; default: "code-rmd").
A path directly to directory "code-rmd" or one of its subdirectories with
R Markdown files specified in input parameter \code{patterns}.
Only a path within a current working workflowr directory is allowed.}

\item{subdirs}{logical (length = 1; default: TRUE).
If \code{FALSE}, file listing will only be directly from a directory in parameter \code{dir_path}.
If \code{TRUE}, file listing will also recurse into subdirectories in parameter \code{dir_path}.}

\item{patterns}{character (length > 0; default: NULL).
If \code{not NULL}, one or more regular expressions that only allow files
with the extension .rmd or .Rmd ("." is also required) is expected.
If \code{NULL}, regular expression "^.*\\.(r|R)md$" is used.}
}
\value{
Nothing if stop rendering is chosen.
Nothing if some check doesn't pass but a stop reason is written and then process stops.
Original or edited 'dir_path' if all checks pass.
}
\description{
Stop processing if rendering of R Markdown files to HTML files isn't possible.
Make small preparations for further processing if needed, like edit \code{dir_path}.
}
