% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-checks.R
\name{initial_checks}
\alias{initial_checks}
\title{Stop process if rendering of .Rmd files into .html files is not possible}
\usage{
initial_checks(dir_path, subdirs, patterns)
}
\arguments{
\item{dir_path}{character of length = 1
Path to a subdirectory, under a main workflowr directory, where .Rmd files for rendering are saved.
A directory name can be first instead of using "./" etc. (e.g. "code-rmd" instead of "./code-rmd").
Examples:
dir_path = "code-rmd"
dir_path = c("code-rmd/subdir1\\subdir2")}

\item{subdirs}{logical of length = 1
If TRUE, file listing will recurse into directories in parameter dir.
If FALSE, file listing will be only directly from directories in parameter dir.}

\item{patterns}{character of length > 0 or NULL
A character vector of paths to .Rmd files for rendering.
If NULL, process all .Rmd files based values in parameters dir and subdirs.
If not NULL, process files matching a regular expression.
Examples:
patterns = "^.\emph{page.}.[  R , r ]md$")
patterns = c("page1.Rmd", ".\emph{page2.}.Rmd")}
}
\value{
FALSE if some check fails, TRUE if all checks were OK,
a logical vector only with TRUE values of length equals to a number of deleted files from directory "analysis"
}
\description{
Evaluate if rendering of .Rmd files into .html files is possible
by checking rules for directories and .Rmd files.
If some check doesn't pass, write a reason and stop processing.
This function is called only from function \code{\link{render_html}} therefore
it's not exported and its input variables have no default values.
}
\examples{
\dontrun{
  initial_checks(dir_path = "code-rmd", subdirs = T, patterns = ".*page.*.(R|r)md$")
}
}
\keyword{subdirectory}
\keyword{workflowr,}
