% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-htmls.R
\name{initial_checks}
\alias{initial_checks}
\title{Stop processing if rendering of .Rmd to .html files is not possible}
\usage{
initial_checks(dir_path, subdirs, patterns)
}
\arguments{
\item{dir_path}{see \code{build_htmls}}

\item{subdirs}{see \code{build_htmls}}

\item{patterns}{see \code{build_htmls}}
}
\value{
Nothing if there's no error but "silent" stop() is needed (if there's no stop() it would return NULL).
Nothing if some check doesn't pass but it writes a reason and stop processing.
A logical vector only with TRUE values of length equals to a number of deleted files from directory "analysis"
if there are files deleted from directory "analysis"
}
\description{
Evaluate if rendering of .Rmd into .html files is possible
by checking rules for directories and .Rmd files.
This function is called only from \code{build_htmls} so its input variables have no default values.
}
