% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_htmls.R
\name{render_to_htmls}
\alias{render_to_htmls}
\title{Render .Rmd to .html files}
\usage{
render_to_htmls(orig_rmd_paths, commit)
}
\arguments{
\item{orig_rmd_paths}{character
Paths to original .Rmd files.}

\item{commit}{see \code{build_htmls}}
}
\value{
Final .html files from their original .Rmd files saved in subdirectories.
}
\description{
Render .Rmd to .html files.
This function is called only from \code{build_htmls} so its input variables have no default values.
}
