% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-htmls.R
\name{render_to_htmls}
\alias{render_to_htmls}
\title{Rendering}
\usage{
render_to_htmls(orig_rmd_paths, commit)
}
\arguments{
\item{orig_rmd_paths}{character (length > 0).
Paths to original R Markdown files.}

\item{commit}{see \code{build_htmls}}
}
\value{
Final HTML files from their original R Markdown files.
}
\description{
Render R Markdown to HTML files.
}
